requirements.txt
certifi==2021.10.8
elastic-transport==8.1.1
elasticsearch==6.8.2
PyYAML==6.0
urllib3==1.26.9
